Added Action calculation
Now the neural network inside creatures brain is working. It calculates the result of each step of the creature based on its sensors.
I will be comenring the code in future for it to be fully understood.
As well as there is plenty of space for speed and coding optimization in that function (calculateCreatureAction).