First working version of simulation
Fixed bug with generation > 1 not having any calculation. Problem was with some lines, that were in wrong if statement.
Simulation does not have mutation implemented yet. It is only copzing genomes to the nex generations.
Most simple way of seeing the simulation is to have safe zone (criteria) on one of the sides of the grid.